requirements.txt
datasets
openai
pandas
python-dotenv
tqdm
pyyaml